git-svn-id: https://clsvn.cienet.com.cn/hzkx-caixiang/code/trunk/Phase3/CXServer_3.0/trunk@66 f47374ae-05c8-11e1-8c5d-33d49f259814

+++ bussiness/src/main/java/com/server/cx/service/cx/impl/UserCommonMGraphicServiceImpl.java
@@ -2,19 +2,13 @@ package com.server.cx.service.cx.impl;
 
 import com.cl.cx.platform.dto.MGraphicDTO;
 import com.google.common.base.Preconditions;
-import com.server.cx.dao.cx.GraphicInfoDao;
-import com.server.cx.dao.cx.HistoryMGraphicDao;
 import com.server.cx.dao.cx.UserCommonMGraphicDao;
 import com.server.cx.dto.OperationResult;
-import com.server.cx.entity.cx.GraphicInfo;
-import com.server.cx.entity.cx.HistoryMGraphic;
 import com.server.cx.entity.cx.UserCommonMGraphic;
-import com.server.cx.exception.CXServerBusinessException;
 import com.server.cx.service.cx.UserCommonMGraphicService;
-import com.server.cx.util.business.AuditStatus;
-import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 /**
  * User: yanjianzou
@@ -23,31 +17,11 @@ import org.springframework.stereotype.Service;
  * FileName:UserCommonMGraphicServiceImpl
  */
 @Service(value = "userCommonMGraphicService")
-public class UserCommonMGraphicServiceImpl extends UserCheckService implements UserCommonMGraphicService {
-
-    @Autowired
-    private GraphicInfoDao graphicInfoDao;
+@Transactional
+public class UserCommonMGraphicServiceImpl extends MGraphicService implements UserCommonMGraphicService {
 
     @Autowired
     private UserCommonMGraphicDao userCommonMGraphicDao;
-
-    @Autowired
-    private HistoryMGraphicDao historyMGraphicDao;
-    private GraphicInfo graphicInfo;
-
-
-
-    private void userCommonMGraphicIdMustBeNotExists(MGraphicDTO mGraphicDTO) {
-        if(mGraphicDTO.getId() != null){
-            throw new CXServerBusinessException("调用一般彩像添加接口错误,Id字段应该为空");
-        }
-    }
-
-    private void userCommonMGraphicIdMustBeExists(MGraphicDTO mGraphicDTO) {
-        if(mGraphicDTO.getId() == null){
-            throw new CXServerBusinessException("调用一般彩像添加接口错误,Id字段应该为空");
-        }
-    }
 
     private void checkAndInitializeContext(String imsi, MGraphicDTO mGraphicDTO) {
         checkParameters(imsi, mGraphicDTO);
@@ -60,6 +34,7 @@ public class UserCommonMGraphicServiceImpl extends UserCheckService implements U
         saveOrUpdateUserCommonMGraphic(mGraphicDTO, userCommonMGraphic);
     }
 
+    @Transactional(readOnly = false)
     private void saveOrUpdateUserCommonMGraphic(MGraphicDTO mGraphicDTO, UserCommonMGraphic userCommonMGraphic) {
         userCommonMGraphic.setActive(true);
         userCommonMGraphic.setName(getGraphicInfoName(mGraphicDTO));
@@ -67,70 +42,42 @@ public class UserCommonMGraphicServiceImpl extends UserCheckService implements U
         userCommonMGraphic.setUserInfo(userInfo);
         userCommonMGraphic.setSignature(mGraphicDTO.getSignature());
         userCommonMGraphicDao.save(userCommonMGraphic);
-    }
-
-    private void historyPreviousUserCommonMGraphic(MGraphicDTO mGraphicDTO) {
-        UserCommonMGraphic previousUserCommonMGraphic = userCommonMGraphicDao.findByUserInfoAndActive(userInfo,true);
-        HistoryMGraphic historyMGraphic = new HistoryMGraphic();
-        historyMGraphic.setGraphicInfo(graphicInfo);
-        historyMGraphic.setUserInfo(userInfo);
-        historyMGraphic.setName(getGraphicInfoName(mGraphicDTO));
-        historyMGraphic.setSignature(mGraphicDTO.getSignature());
-        historyMGraphicDao.save(historyMGraphic);
-        userCommonMGraphicDao.delete(previousUserCommonMGraphic);
-    }
-
-    private String getGraphicInfoName(MGraphicDTO mGraphicDTO) {
-        return StringUtils.isEmpty(mGraphicDTO.getName())? graphicInfo.getName():mGraphicDTO.getName();
-    }
-
-    private void checkParameters(String imsi, MGraphicDTO mGraphicDTO) {
-        Preconditions.checkNotNull(imsi, "imsi编号为空");
-        Preconditions.checkNotNull(mGraphicDTO,"传输对象为空");
-
-        if(mGraphicDTO.getPhoneNos()!= null && mGraphicDTO.getPhoneNos().size()>0){
-            throw new CXServerBusinessException("调用一般彩像添加接口错误,不应该设定号码");
-        }
-
-
-
-    }
-
-    private void getAndCheckGraphicInfo(MGraphicDTO mGraphicDTO) {
-        graphicInfo  = graphicInfoDao.findOne(mGraphicDTO.getId());
-        Preconditions.checkNotNull(graphicInfo, "图像对象不存在");
-        if(!AuditStatus.PASSED.equals(graphicInfo.getAuditStatus())){
-            throw new CXServerBusinessException("图像审核中,或审核未通过");
-        }
+        graphicInfoService.updateGraphicInfoUseCount(graphicInfo);
     }
 
     @Override
     public OperationResult createUserCommonMGraphic(String imsi, MGraphicDTO mGraphicDTO) throws RuntimeException{
         checkAndInitializeContext(imsi, mGraphicDTO);
-        userCommonMGraphicIdMustBeNotExists(mGraphicDTO);
-
-        historyPreviousUserCommonMGraphic(mGraphicDTO);
+        mGraphicIdMustBeNotExists(mGraphicDTO);
+        UserCommonMGraphic previousUserCommonMGraphic = userCommonMGraphicDao.findByUserInfoAndActive(userInfo,true);
+        if(previousUserCommonMGraphic != null){
+            historyPreviousUserCommonMGraphic(previousUserCommonMGraphic);
+            userCommonMGraphicDao.delete(previousUserCommonMGraphic);
+        }
         createAndSaveNewUserCommonMGraphic(mGraphicDTO);
-
         return new OperationResult("createUserCommonMGraphic","success");
     }
 
     @Override
     public OperationResult editUserCommonMGraphic(String imsi, MGraphicDTO mGraphicDTO) {
         checkAndInitializeContext(imsi, mGraphicDTO);
-        userCommonMGraphicIdMustBeExists(mGraphicDTO);
+        mGraphicIdMustBeExists(mGraphicDTO);
         UserCommonMGraphic userCommonMGraphic = userCommonMGraphicDao.findOne(mGraphicDTO.getId());
         saveOrUpdateUserCommonMGraphic(mGraphicDTO,userCommonMGraphic);
         checkAndSetUserInfoExists(imsi);
-
         return new OperationResult("editUserCommonMGraphic","success");
     }
 
     @Override
-    public OperationResult disableUserCommonMGraphic(String imsi, MGraphicDTO mGraphicDTO) {
-        checkAndInitializeContext(imsi,mGraphicDTO);
-        userCommonMGraphicIdMustBeExists(mGraphicDTO);
-        historyPreviousUserCommonMGraphic(mGraphicDTO);
+    public OperationResult disableUserCommonMGraphic(String imsi, String userCommonMGraphicId) {
+        Preconditions.checkNotNull(imsi,"imsi为空");
+        Preconditions.checkNotNull(userCommonMGraphicId,"指定对象不存在");
+        checkAndSetUserInfoExists(imsi);
+        UserCommonMGraphic userCommonMGraphic = userCommonMGraphicDao.findOne(userCommonMGraphicId);
+        if(userCommonMGraphic != null){
+            historyPreviousUserCommonMGraphic(userCommonMGraphic);
+            userCommonMGraphicDao.delete(userCommonMGraphic);
+        }
         return new OperationResult("disableUserCommonMGraphic","success");
     }
 }
